bb-helper: reader_progress: Add integration tests
- Basically moving unit tests to integration tests since they were only
  touching the public API anyway.
